Load the PostHog telemetry provider asynchronously.
`TelemetryProvider` imported `usePostHog` from `posthog-js/react`, whose ESM
entry imports `posthog-js`. That put the library in `LoggedInPage`'s chunk
dependencies, so every logged-in user downloaded 274 KiB of PostHog even with
telemetry disabled.

Extract the PostHog-specific provider into
`logic/telemetry/posthog/PostHogTelemetryProvider` and pull it in via
`loadAsync`, behind the existing `AppConfig.telemetry().enabled` check.

`LoggedInPage` now loads as `e("851e64cc-689")` alone rather than
`Promise.all([e("851e64cc-434"), e("851e64cc-689")])`, so the PostHog chunk is
only fetched when telemetry is actually enabled.
