+++ graylog2-web-interface/src/logic/telemetry/TelemetryProvider.tsx
@@ -14,146 +14,14 @@
  * along with this program. If not, see
  * <http://www.mongodb.com/licensing/server-side-public-license>.
  */
-import React, { useMemo, useEffect, useState, useCallback } from 'react';
-import { usePostHog } from 'posthog-js/react';
-import { useTheme } from 'styled-components';
+import * as React from 'react';
 
-import { getPathnameWithoutId } from 'util/URLUtils';
-import type { TelemetryEventType, TelemetryEvent } from 'logic/telemetry/TelemetryContext';
-import TelemetryContext from 'logic/telemetry/TelemetryContext';
-import { useUpdateTelemetrySettings } from 'logic/telemetry/useTelemetrySettings';
-import TelemetryInfoModal from 'logic/telemetry/TelemetryInfoModal';
-import type { TelemetryDataType } from 'logic/telemetry/useTelemetryData';
-import useTelemetryData from 'logic/telemetry/useTelemetryData';
 import AppConfig from 'util/AppConfig';
-import useSystemDetails from 'hooks/useSystemDetails';
+import loadAsync from 'routing/loadAsync';
 
-const getGlobalProps = (telemetryData: TelemetryDataType) => {
-  const {
-    cluster: {
-      cluster_id,
-      cluster_creation_date,
-      nodes_count,
-      traffic_last_month,
-      users_count,
-      license_count,
-      node_leader_app_version,
-      installation_source,
-    },
-    license,
-    current_user,
-  } = telemetryData;
-
-  return {
-    cluster_id,
-    cluster_creation_date,
-    installation_source,
-    nodes_count,
-    traffic_last_month,
-    users_count,
-    license_count,
-    node_leader_app_version,
-    ...license,
-    ...current_user,
-  };
-};
-
-const PostHogTelemetryProvider = ({ children }: { children: React.ReactElement }) => {
-  const posthog = usePostHog();
-  const theme = useTheme();
-  const system = useSystemDetails();
-
-  const isPosthogLoaded = useCallback(() => posthog?.__loaded === true, [posthog]);
-
-  const { data: telemetryData, isSuccess: isTelemetryDataLoaded, refetch: refetchTelemetryData } = useTelemetryData();
-  const { mutateAsync: updateSettings } = useUpdateTelemetrySettings();
-  const [showTelemetryInfo, setShowTelemetryInfo] = useState<boolean>(false);
-  const [globalProps, setGlobalProps] = useState(undefined);
-
-  useEffect(() => {
-    const app_pathname = getPathnameWithoutId(window.location.pathname);
-    const setGroup = () => {
-      if (isTelemetryDataLoaded && telemetryData?.user_telemetry_settings?.telemetry_enabled) {
-        const {
-          cluster: { cluster_id: clusterId, ...clusterDetails },
-          current_user: { user },
-          license,
-          plugin,
-          search_cluster: searchCluster,
-          data_nodes: dataNodes,
-          user_telemetry_settings: { telemetry_permission_asked: isPermissionAsked },
-        } = telemetryData as TelemetryDataType;
-        setGlobalProps(getGlobalProps(telemetryData));
-
-        posthog.group('cluster', clusterId, {
-          app_pathname,
-          app_theme: theme.mode,
-          ...clusterDetails,
-          ...license,
-          ...plugin,
-          ...searchCluster,
-          ...dataNodes,
-          ...getGlobalProps(telemetryData),
-        });
-
-        posthog.identify(user, { ...getGlobalProps(telemetryData) });
-        setShowTelemetryInfo(!isPermissionAsked);
-      }
-    };
-
-    if (isPosthogLoaded()) {
-      setGroup();
-    }
-  }, [posthog, isTelemetryDataLoaded, telemetryData, theme.mode, isPosthogLoaded]);
-
-  const TelemetryContextValue = useMemo(() => {
-    const sendTelemetry = (eventType: TelemetryEventType, event: TelemetryEvent) => {
-      if (isPosthogLoaded() && globalProps) {
-        try {
-          posthog.capture(eventType, {
-            ...event,
-            ...globalProps,
-            app_theme: theme.mode,
-          });
-        } catch (e) {
-          // eslint-disable-next-line no-console
-          console.warn('Could not capture telemetry event: ', e);
-        }
-      }
-    };
-
-    const sendErrorReport = (error: unknown) => {
-      try {
-        posthog.captureException(error, { version: system?.version });
-      } catch (e) {
-        // eslint-disable-next-line no-console
-        console.warn('Unable to report exception: ', e);
-      }
-    };
-
-    return {
-      sendTelemetry,
-      sendErrorReport,
-    };
-  }, [globalProps, isPosthogLoaded, posthog, theme.mode, system]);
-
-  const handleConfirmTelemetryDialog = () => {
-    updateSettings({ telemetry_permission_asked: true, telemetry_enabled: true }).then(() => {
-      refetchTelemetryData();
-    });
-
-    setShowTelemetryInfo(false);
-  };
-
-  return (
-    <TelemetryContext.Provider value={TelemetryContextValue}>
-      {children}
-      {showTelemetryInfo && (
-        <TelemetryInfoModal show={showTelemetryInfo} onConfirm={() => handleConfirmTelemetryDialog()} />
-      )}
-    </TelemetryContext.Provider>
-  );
-};
+const PostHogTelemetryProvider = loadAsync(
+  () => import(/* webpackChunkName: "PostHogTelemetryProvider" */ 'logic/telemetry/posthog/PostHogTelemetryProvider'),
+);
 
 const isTelemetryEnabled = AppConfig?.telemetry()?.enabled;
 const TelemetryProvider = ({ children }) =>
